+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Change Log
 
 ### 1.1.5
-Make the firmware update message stand out more so that it is not missed
+Make the firmware update message stand out more so that it is not missed.
+Fix a bug when searching for asset documentation with directories that do not have read/search permission.
 
 ### 1.1.4
 Fixed a bug with the recent ModusToolbox applications page