+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+### 1.1.6
+Added command to turn off showing code example README file.
+
 ### 1.1.5
 Make the firmware update message stand out more so that it is not missed.
 Fix a bug when searching for asset documentation with directories that do not have read/search permission.